mudança de dados
