+++ logs/models.csv
@@ -29,3 +29,6 @@ AutoEncoder-2.2-64x64,943671.0,0.519389046555824,21.0,14.0,1.0,0.0,0.0,0.0,0.0,0
 AutoEncoder-2.3-64x64,637501.0,0.4707249385453811,21.0,14.0,1.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,"{'kernel': {'GlorotUniform': 14}, 'bias': {'Zeros': 14}}",0.0,0.0,0.0,0.0
 Conv-1.2-64x64,5.0,0.5,2.0,1.0,1.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,"{'kernel': {'GlorotUniform': 1}, 'bias': {'Zeros': 1}}",0.0,0.0,0.0,0.0
 Conv-1.3-64x64,26.0,0.5,2.0,1.0,1.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,"{'kernel': {'GlorotUniform': 1}, 'bias': {'Zeros': 1}}",0.0,0.0,0.0,0.0
+Unet2.2-64x64,1778481.0,0.2619853439658557,46.0,36.0,4.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,"{'kernel': {'GlorotUniform': 36}, 'bias': {'Zeros': 36}}",0.0,0.0,0.0,0.0
+AutoEncoder-0.1-64x64,919.0,0.4231048240841494,6.0,3.0,1.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,"{'kernel': {'GlorotUniform': 3}, 'bias': {'Zeros': 3}}",0.0,0.0,0.0,0.0
+Conv-12.0-64x64,91791.0,0.4922646178651669,13.0,12.0,1.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,"{'kernel': {'GlorotUniform': 12}, 'bias': {'Zeros': 12}}",0.0,0.0,0.0,0.0
